[IMP] Refactor for files and path. Separated product and account sections from intrastat code.
