+++ l10n_it_intrastat/models/__init__.py
@@ -20,4 +20,6 @@
 ##############################################################################
 
 
+from . import product
+from . import account
 from . import intrastat
